+++ backend/app/controllers/auth/registrations_controller.rb
@@ -1,0 +1,9 @@
+module Auth
+  class RegistrationsController < DeviseTokenAuth::RegistrationsController
+    private
+
+    def sign_up_params
+      params.permit(:email, :password, :password_confirmation, :name)
+    end
+  end
+end
