+++ backend/app/controllers/auth/sessions_controller.rb
@@ -1,0 +1,12 @@
+module Auth
+  class SessionsController < ApplicationController
+    def index
+      if current_user
+        render json: { is_login: true, data: current_user }
+      else
+        render json: { is_login: false, message: 'ユーザーが存在しません' }
+      end
+    end
+  end
+end
+
