+++ backend/config/routes.rb
@@ -5,7 +5,7 @@ Rails.application.routes.draw do
   post "/graphql", to: "graphql#execute"
 
   mount_devise_token_auth_for 'User', at: 'auth', controllers: {
-    registrations: 'api/auth/registrations'
+    registrations: 'auth/registrations'
   }
 
   namespace :auth do
